First upload - description and instructions
V2.1.1

A wifi jammer for use on 2.4Ghz networks, jamming is done on 3 ranges, from 2412Mhz to 2483Mhz.
Intended to work with a 20x4 lcd screen, and a status LED.
The jammer detects the active channels in your surrounding  on startup, and displays them before the initial run or after
Running was stopped.
Running continuous on the selected frequencies range until a button was pressed. Upon a second button press, the corresponding
Range will be jammed.